[Update] update extension
