+++ QRArtGenerator/Common/CommonShape.swift
@@ -1,0 +1,19 @@
+//
+//  CommonShape.swift
+//  QRArtGenerator
+//
+//  Created by khac tao on 30/06/2023.
+//
+
+import Foundation
+import SwiftUI
+
+struct RoundedCorner: Shape {
+    var radius: CGFloat = .infinity
+    var corners: UIRectCorner = .allCorners
+    func path(in rect: CGRect) -> Path {
+        let path = UIBezierPath(roundedRect: rect, byRoundingCorners: corners, cornerRadii: CGSize(width: radius, height: radius))
+        return Path(path.cgPath)
+    }
+}
+
