+++ QRArtGenerator/Helper/AppHelper.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import SwiftUI
 
 struct AppHelper {
     
@@ -14,3 +15,13 @@ struct AppHelper {
     }
     
 }
+
+struct SafeAreaInsetsKey: EnvironmentKey {
+    static var defaultValue: EdgeInsets {
+        (UIApplication.shared.connectedScenes
+            .compactMap { $0 as? UIWindowScene }
+            .flatMap { $0.windows }
+            .first(where: { $0.isKeyWindow })?.safeAreaInsets ?? .zero).insets
+    }
+}
+
